+++ system/src/Grav/Framework/Psr7/Request.php
@@ -1,0 +1,3 @@
+<?php
+
+class_alias('Nyholm\Psr7\Request', 'Grav\Framework\Psr7\Request');