+++ system/src/Grav/Framework/Psr7/Response.php
@@ -1,0 +1,3 @@
+<?php
+
+class_alias('Nyholm\Psr7\Response', 'Grav\Framework\Psr7\Response');