+++ system/src/Grav/Framework/Psr7/ServerRequest.php
@@ -1,0 +1,3 @@
+<?php
+
+class_alias('Nyholm\Psr7\ServerRequest', 'Grav\Framework\Psr7\ServerRequest');