+++ system/src/Grav/Framework/Psr7/Stream.php
@@ -1,0 +1,3 @@
+<?php
+
+class_alias('Nyholm\Psr7\Stream', 'Grav\Framework\Psr7\Stream');