+++ system/src/Grav/Framework/Psr7/UploadedFile.php
@@ -1,0 +1,3 @@
+<?php
+
+class_alias('Nyholm\Psr7\UploadedFile', 'Grav\Framework\Psr7\UploadedFile');